requirements.txt
scanpy>=1.9.0
anndata>=0.8.0
cell2location>=0.1.0
scvi-tools>=0.16.0
pandas>=1.4.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.12.0
